[CSHW-4]: Домашнее задание 4

See merge request natasha.buranova/Route256Five!4

+++ src/Ozon.Route256.Five.OrderService/Controllers/DTO/Orders/OrdersInRegionResponse.cs
@@ -4,7 +4,7 @@ public class OrdersInRegionResponse
 {
     public string? RegionName { get; set; }
     public int CountOrders { get; set; }
-    public long TotalSumOrders { get; set; }
+    public double TotalSumOrders { get; set; }
     public long TotalWeight { get; set; }
     public long CountClients { get; set; }
 }
